+++ docs/source/changelog.rst
@@ -1,13 +1,20 @@
 Changelog
 =========
 
-2.3.6 (unreleased)
-------------------
+2.4 (unreleased)
+----------------
 
-* Add :py:meth:`Query.with_node_kinds <liara.query.Query.with_node_kinds>` and :py:meth:`Query.without_node_kinds <liara.query.Query.without_node_kinds>` to :py:class:`~liara.query.Query`. This allows lists of nodes (as returned by :py:meth:`~liara.template.SiteTemplateProxy.select` and other functions) to be filtered by node kind. This is useful if you want to mix static content and documents in the same folder.
+.. note::
+
+  This release contains breaking changes when using indices and collections. Make sure to review the changelog and the :doc:`collections` and :doc:`indices` pages when running into issues after updating.
+
+  The main change is that missing metadata fields in collections and indices result in an error now, instead of silently removing items. Use ``filter_by`` to filter nodes missing specific metadata fields. When updating from < 2.4, you can simply copy the ``order_by``/``group_by`` entry in the respective YAML file into ``filter_by`` to get the original behavior back.
+
+* Add :py:meth:`Query.with_node_kinds <liara.query.Query.with_node_kinds>` and :py:meth:`Query.without_node_kinds <liara.query.Query.without_node_kinds>` to :py:class:`~liara.query.Query`. This allows lists of nodes (as returned by :py:meth:`~liara.template.SiteTemplateProxy.select` and other functions) to be filtered by the node kind. This is useful if you want to mix static content and documents in the same folder.
 * Add :py:meth:`SiteTemplateProxy.select_pages <liara.template.SiteTemplateProxy.select_pages>` and :py:attr:`Page.children <liara.template.Page.children>` to select pages (i.e. document and index nodes) only without having to manually filter the result using ``with_node_kinds``/``without_node_kinds``.
 * Add ``node_kinds`` to :py:meth:`Collection.__init__ <liara.site.Collection.__init__>` to allow constraining a collection to a specific node kind. See :doc:`collections` for more details.
 * Improve the debug output during publishing. The template publisher will now print which document is published using which template. As part of this change, :py:attr:`Template.path <liara.template.Template.path>` was added.
+* Change how :py:meth:`Page.references <liara.template.Page.references>` is populated for *top level indices*. Previously, ``references`` would not be populated for a top-level index. Additionally, improve the documentation of top-level indices, see :doc:`indices` for more details.
 * Improve error handling when trying to sort nodes which are missing the corresponding metadata key, for example, using :py:meth:`~liara.query.Query.sorted_by_title`. Previously, this would raise an exception about a failed comparison involving ``None``, now this raises a more useful exception which contains the path to the item missing the metadata key and which key was requested.
 
 2.3.5